Working Solution but unsatisfactory.
This logic will beat Ctrl & Sustain but it is not ideal. the logic is broken, it crafts a lot of ethers which wins the game but the logic is not creating them for the right reasons. essentially this is bugged code that happens to win the game.
Plus this code causes a loss on "stat. eff. ctrl." and "lmt brk crafter".