Working solution for Ctrl & Sustain and other fixes
added turtle mode logic for the team that only applies in the Ctrl & Sustain level.

fixed one shot logic
added regions
moved some of the cleric's brave advance logic to the top of the function.
some other small logic changes?

NOTE
changes in this team now cause item crafter to level to fail and others to take long to beat. I attribute this to me fixing the one shot code. the team was unintentionally more aggressive wit the broken code.